+++ src/LPPaver/Constraint/Util.hs
@@ -1,3 +1,12 @@
+{-|
+Module      : LPPaver.Constraint.Util
+Description : Utility function for the 'Constraint' type
+Copyright   : (c) Junaid Rasheed, 2021-2022
+License     : MPL
+Maintainer  : jrasheed178@gmail.com
+Stability   : experimental
+Module defining useful utility functions for the 'Constraint' data type.
+-}
 {-# LANGUAGE LambdaCase #-}
 
 module LPPaver.Constraint.Util where
@@ -8,11 +17,19 @@ import Data.List (nub)
 import qualified Linear.Simplex.Types as LT
 import qualified Data.Map as M
 import Data.Maybe
+import Linear.Simplex.Types (PolyConstraint)
 
+-- | Get the LHS of a 'Constraint'
 constraintLeftSide :: Constraint -> [(String, Rational)]
 constraintLeftSide (GEQ lhs _) = lhs
 constraintLeftSide (LEQ lhs _) = lhs
 
+-- | Get the LHS of a 'Constraint'
+constraintRightSide :: Constraint -> Rational
+constraintRightSide (GEQ _ rhs) = rhs
+constraintRightSide (LEQ _ rhs) = rhs
+
+-- | Get a list of all variables in a list of 'Constraint's
 constraintVars :: [Constraint] -> [String]
 constraintVars cs = nub $ aux cs
   where
@@ -20,6 +37,8 @@ constraintVars cs = nub $ aux cs
     aux [] = []
     aux (x : xs) = map fst (constraintLeftSide x) ++ aux xs
 
+-- | Convert a list of 'Constraint's into a list of 'LT.PolyConstraint's.
+-- The function returns a pair with the first item being a list of 'LT.PolyConstraint's and the second item being a mapping of 'String' variables (from 'Constraint's) to 'Integer' variables (for 'LT.PolyConstraint's).
 constraintsToSimplexConstraints :: [Constraint] -> ([LT.PolyConstraint], M.Map String Integer)
 constraintsToSimplexConstraints constraints =
   (